Move styles to separate file
